pyproject.toml
# ─────────────────────────────────────────────────────────────────────────────
# pyproject.toml
# Modern Python packaging configuration (PEP 517/518).
# Replaces setup.py for new projects.
#
# Install the package in editable/development mode:
#   pip install -e .
# ─────────────────────────────────────────────────────────────────────────────

[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# ── Package metadata ──────────────────────────────────────────────────────────
[project]
name            = "research-literature-agent"
version         = "1.0.0"
description     = "AI agent that automatically generates literature reviews from arXiv papers"
readme          = "README.md"
license         = { file = "LICENSE" }
requires-python = ">=3.10"

authors = [
    { name = "Your Name", email = "your.email@example.com" }
]

keywords = [
    "literature-review", "arxiv", "langchain", "llm",
    "research-automation", "nlp", "ai-agent"
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Runtime dependencies (mirrors requirements.txt)
dependencies = [
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.30.0",
    "arxiv>=2.1.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage      = "https://github.com/YOUR_USERNAME/research-literature-agent"
Repository    = "https://github.com/YOUR_USERNAME/research-literature-agent"
Issues        = "https://github.com/YOUR_USERNAME/research-literature-agent/issues"
Documentation = "https://github.com/YOUR_USERNAME/research-literature-agent#readme"

# ── CLI entry point ───────────────────────────────────────────────────────────
# After `pip install -e .`, you can run: research-agent --topic "..."
[project.scripts]
research-agent = "research_agent.agent:_build_cli_parser"

# ── Tool configurations ───────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"
# Exclude integration tests by default (they make real API calls)
# Run them explicitly with: pytest -m integration
markers = [
    "integration: marks tests that call real APIs (deselect with '-m not integration')",
    "slow: marks tests that take more than 5 seconds",
]

[tool.coverage.run]
source   = ["research_agent"]
omit     = ["tests/*", "examples/*"]

[tool.coverage.report]
show_missing = true
fail_under   = 70      # fail CI if coverage drops below 70%

[tool.black]
line-length    = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in default arguments
]

[tool.mypy]
python_version         = "3.10"
warn_return_any        = true
warn_unused_configs    = true
ignore_missing_imports = true   # many LangChain stubs are incomplete
